refactor(module-health): split gitlab api/client and scanner factories by concern, scrub collaborator handle from test fixtures
