+++ tests/teatree_core/test_author_trust.py
@@ -159,12 +159,14 @@ class TestExtraTrustedUnion(TestCase):
     def test_extra_trusted_defaults_to_empty_so_existing_callers_are_unchanged(self) -> None:
         _seed_known()
         with _public():
-            assert classify_author(_PUBLIC, "adrien-oper").untrusted is True
+            assert classify_author(_PUBLIC, "trusted-colleague").untrusted is True
 
     def test_extra_trusted_handle_is_trusted(self) -> None:
         _seed_known()
         with _public():
-            classification = classify_author(_PUBLIC, "adrien-oper", extra_trusted=frozenset({"adrien-oper"}))
+            classification = classify_author(
+                _PUBLIC, "trusted-colleague", extra_trusted=frozenset({"trusted-colleague"})
+            )
         assert classification.trusted is True
         assert classification.untrusted is False
 
@@ -172,13 +174,13 @@ class TestExtraTrustedUnion(TestCase):
         """A DB-trusted handle stays trusted when the caller supplies its own extras."""
         _seed_known()
         with _public():
-            assert classify_author(_PUBLIC, "souliane", extra_trusted=frozenset({"adrien-oper"})).trusted
+            assert classify_author(_PUBLIC, "souliane", extra_trusted=frozenset({"trusted-colleague"})).trusted
 
     def test_extra_trusted_match_is_case_insensitive(self) -> None:
-        assert is_trusted_author("Adrien-Oper", extra_trusted=frozenset({"adrien-oper"})) is True
+        assert is_trusted_author("Trusted-Colleague", extra_trusted=frozenset({"trusted-colleague"})) is True
 
     def test_extra_trusted_never_admits_an_unlisted_author(self) -> None:
-        assert is_trusted_author("evilhacker", extra_trusted=frozenset({"adrien-oper"})) is False
+        assert is_trusted_author("evilhacker", extra_trusted=frozenset({"trusted-colleague"})) is False
 
     def test_empty_author_is_untrusted_even_with_extras(self) -> None:
-        assert is_trusted_author("", extra_trusted=frozenset({"adrien-oper"})) is False
+        assert is_trusted_author("", extra_trusted=frozenset({"trusted-colleague"})) is False
